Replace shrinking algorithm with simpler one
The shrinking algorithm for command lists actually ends up being less
efficient, at least for situations where a lot of the
commands (proportionally) are necessary for reproducing the issue.

I've replaced it with the shrinking that is used for vectors in
test.check more generally, and I guess it'll be a space for future
experimentation if I ever feel like it.