Clean up code and add docstrings
No functionality should have changed in any noticable way, but the code
is a bit cleaner in many ways. Slight downside is that now the output of
quick-check on a failure is a bit less readable. Ideally, though, the
`print-test-results` will be used to interpret that (which does a much
better job).