Extract command-utils, add docstrings
I've extracted some functions to a new command-utils namespace in order
to simplify all the (or (:model/next-state command) ...) sort of stuff.
It was getting a bit too verbose for my liking.

Docstrings have been added in more places, just to give a bit of an idea
about what things are doing. They're not particularly well written, but
they're more useful than nothing.