Separate "current command" and "command list"
In the command runner/varifier there was an implicit "the current
command is the first one in the command list". This seemed a bit of a
problem to me, particularly because we modified it as the first
operation when a command became "current". In order to fix this, the
current command is now its own entry in the runner states.