Move functionality from core to core-utils
The basic idea with this is that we'd like to be able to say
"use (:require '[stateful-test.core :refer [specification-correct?]])"
and have them be able to, from that alone, use stateful-check. With the
movement of things into core-utils, though, people can't accidentally
pull something incorrect out.

The two deprecated functions (print-test-results and
reality-matches-model) are both still in core, too, as they're needed so
as to not break things too much.