Remove ^:private from everywhere
The namespaces are sufficiently broken up now that there's really no
need to have the ^:private metadata on things. When stateful-check.core
was a big namespace I wanted to have things be private so they wouldn't
be accidentally referred into other namespaces. Now stateful-check.core
is much smaller, so that's not really a concern for me.