Make :model/args results be coerced to a generator
The result of :model/args can be interpreted a little bit intelligently.
This change makes it do some common-sense things.

If you return a generator: the generator will be used as-is.

If you return a vector (or other sequential collection): each element
will be coerced into a generator, then gen/tuple will be called with all
of the elements.

If you return a map: each value will be coerced into a generator, then
gen/hash-map will be called with all the key/value-gen pairs.

If you return anything else: gen/return will be called with the value.