Simplify runner, add :model/requires to verifier
The command runner currently checks preconditions, which isn't actually
necessary (the preconditions should be checked by the verifier already).

After adding the :model/requires key to the command map I forgot to add
it to the verifier, which meant that shrinking was coming up with
illegal test cases. That's fixed now!