Make command printing print stuff before mutations
When printing the results of commands we had a bit of a problem where
things were printed with their state taken from the end of the sequence
of commands rather than at the point where they happened.

For example, imagine a test for a set:
  #<0> = (:new) => #{10}
  #<1> = (:contains? #<0> 10) => false
  #<2> = (:add #<0> 10) => true
  #<3> = (:contains? #<0> 10) => true

This is clearly wrong: when the set was created it had zero elements.
The 10 was only added in later. At the end of the test, though, the set
had the 10 in it, so it was printed as such. This is unhelpful.

Now, it would print like this:
  #<0> = (:new) => #{}
  #<1> = (:contains? #<0> 10) => false
  #<2> = (:add #<0> 10) => true
  #<3> = (:contains? #<0> 10) => true

This is more accurate, and much better for debugging.