Allow number of tries per test to be specified
The number of tries determines how many times the same series of
commands should be executed in order to get a result. If any one of the
attemps fails then the test would be considered failed. This is useful
for non-deterministic tests.

As a result of this, we can't just re-run commands to print intermediate
results, and so we now capture the actual failing test case (and
results) in order to construct the output. This has the problem of
re-introducing the issue where printed results are incorrect when
printed after the run (due to mutations and the like). In order to fix
this we need to convert things to strings during execution, somehow.