Shrink more aggressively
Shrinking is now much more aggressive than it used to be. This can
result in shrinking taking longer than it used to, but it also means
that shrinking can yield much better results.

It used to just try to remove each command, one by one, and then attempt
to shrink the arguments to each command, one by one.

It now tries to remove either a chunk of adjacent commands, or a pair of
commands (which may not be adjacent). It then tries to shrink command
arguments, then tries a hybrid (shrink a command and remove a chunk/pair
at the same time).

This means it can reduce a queue example like:

  (:new 2)
  (:push 0)
  (:push 0)
  (:pop)
  (:push 0)
  (:count) => -1

Into something more like:

  (:new 1)
  (:push 0)
  (:pop)
  (:push 0)
  (:count) => -1

See test/stateful_check/java_queue_test.clj for an example of this.