Only use custom printing logic for our exceptions (HACK)
This is a hack, but it's necessary for proper reporting when an
exception is thrown outside of the "tries" loop. Most notably this
happens when a command doesn't have a :real/command key (because we
throw an AssertionError). Ideally we'd have our own custom exception
type rather than reading the message, but that's surprisingly hard in
Clojure.