Warn a user if a failure appears to be due to mutation
The return value of :command functions should be immutable if it's
going to be checked with a postcondition. We can try to verify that by
comparing pr-str outputs when the command ran, and at the end of the
execution. Then we can give the user at least a bit of a heads up.
This will fail easily if you're trying to make it fail, but it should
help if an object's pr-str output contains something of its internal
state.