Remove the fancy length logic
I tried to scale the max-length based on size, and where we're up to
relative to our max-size, but it ended up being a bad idea. The logic
wasn't too hard, but it led to tests finding less bugs because they
tended to be too short to do anything meaningful.