Fix the tests, give them more chance of passing
The problem with testing the framework itself is that the randomness
comes back to bite you a bit. Particular for the race condition
things. Usually with this stuff you can't have a false positive (i.e.
a test fail when it should have passed), but then when you negate it
then everything flips. Many of my tests often fail, even though they
are correct. Usually running them a few times is enough, but tweaking
the numbers can help too.