Fix tables in specification.org
This is a weird bug. Emacs was not displaying the bytes in the file in
a normal way. If I used "cat" to view the file it worked fine, but if
I used Emacs it rendered some characters as invisible (or, really, I
think as a negative width character, somehow). It's all very
confusing. Hopefully it should work now.